Android: toolchain-agnostic dependency lookup and an Android preset
The build found every C library with pkg_check_modules against the build machine's pkg-config, which cannot see a
cross-compiled sysroot. qt/cmake/XojDeps.cmake now makes one target, xoj::deps: libxml2, libzip and qpdf as CMake
packages first where the toolchain provides them (vcpkg), pkg-config for the GNOME libraries, which only ship .pc
files (vcpkg installs those too). A cross build through vcpkg reads only the target triplet's .pc files, and static
dependencies (XQT_STATIC_DEPS, on for Android) use pkg-config's --static link line so that their private
dependencies are linked. The Linux desktop stays on pkg-config alone: Ubuntu 22.04's libzip CMake files are broken
and fail even a QUIET find_package. Its link line is unchanged (compared with the main checkout's build).

qt/CMakePresets.json has android-arm64-debug: Qt's Android toolchain chaining vcpkg's, which chains the NDK's, as
QField does. On Android the tests, the CLI with the golden tests, the spikes and the .deb packaging are off by
default (XQT_BUILD_TESTS/CLI/SPIKES; Qt Test is only looked for when tests are built), and the spikes' own cairo and
glib lookups moved into their block.

Two things the NDK's libc++ does not have: floating point std::from_chars (checked now, as upstream's CMake does;
the core then falls back to g_ascii_strtod) and std::to_string of a file time count, which is __int128 there
(HybridPdf, PdfPageKeeper, PageClipboard now cast to long long, the same value on Linux).

Desktop: build-qt configures and builds; ctest -L unit (130), -L session (68), -L shell (196) and golden-quick pass.

+++ qt/src/shell/PageClipboard.cpp
@@ -23,7 +23,7 @@ std::string stampOf(const fs::path& p) {
         return {};
     }
     const auto time = fs::last_write_time(p, ec);
-    return std::to_string(size) + ":" + std::to_string(time.time_since_epoch().count());
+    return std::to_string(size) + ":" + std::to_string(static_cast<long long>(time.time_since_epoch().count()));
 }
 }  // namespace
